Remove labels from Encoder keys for now and render correctly in test tab
